requirements.txt
fastapi
sqlalchemy
alembic
mysql-connector-python
uvicorn
pytest
httpx
python-jose
python-dotenv